[test] merge symbolic shape codegen regression tests into alloc_codegen (#1301)
